KAN-85: Rename: move to signaling package
